requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
colorlog==6.8.2
cvzone==1.6.1
Django==5.0.6
django-environ==0.11.2
django-simpleui==2024.4.1
djangorestframework==3.15.1
dlib==19.19.0
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.7
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
requests==2.32.2
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
